requirements.txt
pandas
openpyxl
PyMuPDF
streamlit
pytesseract
pdf2image
Pillow